refactor: Claim screen
